+++ app/Console/Commands/CleanupStatisticsCommand.php
@@ -31,13 +31,17 @@ class CleanupStatisticsCommand extends Command
     public function handle()
     {
         // Remove all server statistics data older than the retention period
-        $serverDay = now()->subDays(setting('statistics.servers.retention_period'))->toDateString();
-        Log::info('Removing server statistics data older than '.$serverDay);
-        ServerStat::where('created_at', '<', $serverDay)->delete();
+        if (setting('statistics.servers.retention_period') != -1) {
+            $serverDay = now()->subDays(setting('statistics.servers.retention_period'))->toDateString();
+            Log::info('Removing server statistics data older than '.$serverDay);
+            ServerStat::where('created_at', '<', $serverDay)->delete();
+        }
 
         // Remove all meeting statistics data older than the retention period
-        $meetingDay = now()->subDays(setting('statistics.meetings.retention_period'))->toDateString();
-        Log::info('Removing meeting statistics data older than '.$serverDay);
-        MeetingStat::where('created_at', '<', $meetingDay)->delete();
+        if (setting('statistics.meetings.retention_period') != -1) {
+            $meetingDay = now()->subDays(setting('statistics.meetings.retention_period'))->toDateString();
+            Log::info('Removing meeting statistics data older than '.$serverDay);
+            MeetingStat::where('created_at', '<', $meetingDay)->delete();
+        }
     }
 }
